pyproject.toml
[tool.poetry]
name = "wedding-list"
version = "0.1.0"
description = ""
authors = ["morfinPL <pl.morfin.pl@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^1.12.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
